Win32 process improved Help menu linked with Pinguino wiki New CDC functions (ex. CDC.printf) Added Zigbee library
Pinguino IDE Beta X.2
8-bit / sdcc 2.9.0  + gputils 0.13.7
32-bit / g++ 4.5.2 + binutils 2.21 + newlib 1.19
No more support for Microchip C32 Compiler.
Multi-platform IDE (one Python file for all OS).
Improved Command Line mode.
Greater Compatibily with Arduino.
exceptions : 
- Ethernet (to do)
- Firmata (to do)
- Matrix (not planned)
- Sprite (not planned)
- SoftwareSerial (not planned)
- String (C++)

+++ branches/x.2/p32/include/pinguino/libm/asinf.c
@@ -28,9 +28,9 @@
 
 #include <math.h>
 
-float asincosf(const float x, const int isacos);
+double asincosf(const double x, const int isacos);
 
-float asinf(const float x) _MATH_REENTRANT
+double asinf(const double x) _MATH_REENTRANT
 {   
          if(x== 1.0) return  HALF_PI;
     else if(x==-1.0) return -HALF_PI;
